fix: 🐛 Fixed "hit" class for nums repeated on both player cards
Got code to simultaneously add "hit" class (and pink circle to square in
HTML) to both cards if class of td matches bingo number called. Changed
tr classes to "rowNum" vs just "Num" to do so, and changed makeTableRow
function to add number shown on square as a class identifier instead of
an id identifier.